requirements.txt
pandas
OpenAI
python-dotenv
matplotlib
streamlit